pyproject.toml
[tool.poetry]
name = "jarwo-lookup"
version = "0.1.0"
description = ""
authors = ["Samwoof <abrisam.rm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^3.0.0"
sentence-transformers = "^2.2.2"
transformers = "^4.36.2"
safetensors = "^0.4.1"
pdf2image = "^1.17.0"
pytesseract = "^0.3.10"
PyMuPDF = "^1.23.18"
tabulate = "^0.9.0"
pandas = "^2.2.0"
chromadb = "^0.4.22"
Flask-SQLAlchemy = "^3.1.1"
PyYAML = "^6.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
